requirements.txt
Django==5.2.7
mysqlclient==2.2.7
python-dotenv==1.1.1
Pillow==11.1.0
django-modeltranslation==0.19.11
httpx==0.28.1